Rename CLI tool from spawn_cli to spawnctl
