--- spawn_cli/lib/spawn_cli/commands/dev.ex
@@ -1,7 +1,0 @@
-defmodule SpawnCli.Commands.Dev do
-  use DoIt.Command,
-    name: "dev",
-    description: "Manages local development."
-
-  subcommand(SpawnCli.Commands.Dev.Run)
-end
