--- spawn_cli/lib/spawn_cli/runtime/adapters/elixir/elixir_runtime.ex
@@ -1,2 +1,0 @@
-defmodule SpawnCli.Runtime.Adapters.Elixir.ElixirRuntime do
-end
