--- spawn_cli/lib/spawn_cli/runtime/adapters/nodejs/nodejs_runtime.ex
@@ -1,2 +1,0 @@
-defmodule SpawnCli.Runtime.Adapters.Nodejs.NodejsRuntime do
-end
