--- spawn_cli/lib/spawn_cli/runtime/adapters/python/python_runtime.ex
@@ -1,2 +1,0 @@
-defmodule SpawnCli.Runtime.Adapters.Python.PythonRuntime do
-end
