--- spawn_cli/lib/spawn_cli/runtime/adapters/spring/spring_runtime.ex
@@ -1,2 +1,0 @@
-defmodule SpawnCli.Runtime.Adapters.Spring.SpringRuntime do
-end
