--- spawn_cli/lib/spawn_cli/runtime/runtime_adapter.ex
@@ -1,2 +1,0 @@
-defmodule SpawnCli.Runtime.RuntimeAdapter do
-end
