--- spawn_cli/test/spawn_cli_test.exs
@@ -1,4 +1,0 @@
-defmodule SpawnCliTest do
-  use ExUnit.Case
-  doctest SpawnCli
-end
