+++ spawnctl/.k8s/example/system.yaml
@@ -1,0 +1,15 @@
+---
+apiVersion: spawn-eigr.io/v1
+kind: ActorSystem
+metadata:
+  name: spawn-system
+  namespace: default
+spec:
+  externalInvocation:
+    enabled: "false"
+    externalConnectorRef: invocation-connection-ref-secret
+  statestore:
+    type: MySql
+    credentialsSecretRef: mysql-connection-secret
+    pool:
+      size: "3"
