+++ spawnctl/lib/spawnctl/runtime/adapters/elixir/elixir_runtime.ex
@@ -1,0 +1,2 @@
+defmodule SpawnCtl.Runtime.Adapters.Elixir.ElixirRuntime do
+end
