+++ spawnctl/lib/spawnctl/runtime/adapters/nodejs/nodejs_runtime.ex
@@ -1,0 +1,2 @@
+defmodule SpawnCtl.Runtime.Adapters.Nodejs.NodejsRuntime do
+end
