+++ spawnctl/lib/spawnctl/runtime/adapters/python/python_runtime.ex
@@ -1,0 +1,2 @@
+defmodule SpawnCtl.Runtime.Adapters.Python.PythonRuntime do
+end
