+++ spawnctl/lib/spawnctl/runtime/adapters/spring/spring_runtime.ex
@@ -1,0 +1,2 @@
+defmodule SpawnCtl.Runtime.Adapters.Spring.SpringRuntime do
+end
