+++ spawnctl/lib/spawnctl/runtime/runtime_adapter.ex
@@ -1,0 +1,2 @@
+defmodule SpawnCtl.Runtime.RuntimeAdapter do
+end
