+++ spawnctl/test/spawn_cli_test.exs
@@ -1,0 +1,4 @@
+defmodule SpawnCliTest do
+  use ExUnit.Case
+  doctest SpawnCli
+end
